Split greedy solver into header + implementation
Implementation path:
core/src/ambiguity_resolution/greedy_ambiguity_resolution_algorithm.cpp

Header path:
core/include/traccc/ambiguity_resolution/greedy_ambiguity_resolution_algorithm.hpp
